% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{document_word_count}
\alias{document_word_count}
\title{Count words in active document}
\usage{
document_word_count()
}
\value{
Number of words
}
\description{
Count the words in the active source editor document.
}
\seealso{
Other word_count: \code{\link{selection_word_count}}
}
