% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{selection_word_count}
\alias{selection_word_count}
\title{Count words in selection}
\usage{
selection_word_count()
}
\value{
Number of words
}
\description{
Count the words in the selected text in the source editor.
}
\seealso{
Other word_count: \code{\link{document_word_count}}
}

